backend/requirements.txt
absl-py==0.12.0
appdirs==1.4.4
asgiref==3.3.1
audioread==2.1.9
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cycler==0.10.0
decorator==4.4.2
Django==3.1.7
django-appconf==1.0.4
django-imagekit==4.0.2
djangorestframework==3.12.2
fontawesome-free==5.15.3
google-auth==1.28.0
google-auth-oauthlib==0.4.4
grpcio==1.36.1
idna==2.10
importlib-metadata==3.10.0
inflect==5.3.0
joblib==1.0.1
kiwisolver==1.3.1
librosa==0.8.0
llvmlite==0.36.0
Markdown==3.3.4
matplotlib==3.4.1
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
mysqlclient==2.0.3
numba==0.53.1
numpy==1.20.2
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.1.48
packaging==20.9
pilkit==2.0
Pillow @ file:///C:/ci/pillow_1615224334542/work
pooch==1.3.0
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.7.2
scikit-learn==0.24.1
scipy==1.6.2
six @ file:///C:/ci/six_1605205426665/work
SoundFile==0.10.3.post1
sqlparse==0.4.1
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.1
threadpoolctl==2.1.0
torch==1.8.1
torchaudio==0.7.0
torchvision==0.9.1
tqdm==4.59.0
typing-extensions @ file:///home/ktietz/src/ci_mi/typing_extensions_1612808209620/work
Unidecode==1.2.0
urllib3==1.26.4
Werkzeug==1.0.1
wincertstore==0.2
yacs==0.1.8
zipp==3.4.1
django-imagekit
